+++ src/main/java/com/smartcampus/erp/dto/AssignmentResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -18,5 +19,5 @@ public class AssignmentResponse {
     private String title;
     private String description;
     private LocalDateTime dueDate;
-    private Double maxMarks;
+    private BigDecimal maxMarks;
 }
