+++ src/main/java/com/smartcampus/erp/entity/AssignmentSubmission.java
@@ -2,6 +2,7 @@ package com.smartcampus.erp.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -32,5 +33,5 @@ public class AssignmentSubmission {
     private LocalDateTime submittedAt;
 
     @Column(name = "marks_obtained")
-    private Double marksObtained;
+    private BigDecimal marksObtained;
 }
